requirements.txt
Django
django-cors-headers
djangorestframework
drf-yasg
djangorestframework-simplejwt